[hdpowerview] preliminary work on generation 3 api

[hdpowerview] refactoring, bug fixes


[hdpowerview] create scene v1 & v3; refactor gson stuff


[hdpowerview] make target classes final


[hdpowerview] cosmetics


[hdpowerview] reverse V3 additions


[hdpowerview] clean folder structure


[hdpowerview] fixed comparable bug


[hdpowerview] prliminary work on generation 3 api


[hdpowerview] more Generation 3 stuff added


[hdpowerview] adapt folder structure


[hdpowerview] refactoring, bug fixes


[hdpowerview] create scene v1 & v3; refactor gson stuff


[hdpowerview] make target classes final


[hdpowerview] cosmetics


[hdpowerview] reverse V3 additions


[hdpowerview] clean folder structure


[hdpowerview] fixed comparable bug
