+++ bundles/org.openhab.binding.hdpowerview/src/main/java/org/openhab/binding/hdpowerview/internal/HDPowerViewWebTargets.java
@@ -12,6 +12,7 @@
  */
 package org.openhab.binding.hdpowerview.internal;
 
+import java.lang.reflect.Type;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.List;
@@ -29,30 +30,18 @@ import org.eclipse.jetty.http.HttpMethod;
 import org.eclipse.jetty.http.HttpStatus;
 import org.openhab.binding.hdpowerview.internal.api.Color;
 import org.openhab.binding.hdpowerview.internal.api.HubFirmware;
+import org.openhab.binding.hdpowerview.internal.api.ShadeData;
 import org.openhab.binding.hdpowerview.internal.api.ShadePosition;
 import org.openhab.binding.hdpowerview.internal.api.SurveyData;
 import org.openhab.binding.hdpowerview.internal.api.UserData;
-import org.openhab.binding.hdpowerview.internal.api.requests.RepeaterBlinking;
-import org.openhab.binding.hdpowerview.internal.api.requests.RepeaterColor;
-import org.openhab.binding.hdpowerview.internal.api.requests.ShadeCalibrate;
-import org.openhab.binding.hdpowerview.internal.api.requests.ShadeJog;
-import org.openhab.binding.hdpowerview.internal.api.requests.ShadeMove;
-import org.openhab.binding.hdpowerview.internal.api.requests.ShadeStop;
-import org.openhab.binding.hdpowerview.internal.api.responses.FirmwareVersion;
-import org.openhab.binding.hdpowerview.internal.api.responses.Repeater;
 import org.openhab.binding.hdpowerview.internal.api.responses.RepeaterData;
 import org.openhab.binding.hdpowerview.internal.api.responses.Repeaters;
+import org.openhab.binding.hdpowerview.internal.api.responses.Scene;
 import org.openhab.binding.hdpowerview.internal.api.responses.SceneCollections;
-import org.openhab.binding.hdpowerview.internal.api.responses.SceneCollections.SceneCollection;
 import org.openhab.binding.hdpowerview.internal.api.responses.Scenes;
-import org.openhab.binding.hdpowerview.internal.api.responses.Scenes.Scene;
+import org.openhab.binding.hdpowerview.internal.api.responses.ScheduledEvent;
 import org.openhab.binding.hdpowerview.internal.api.responses.ScheduledEvents;
-import org.openhab.binding.hdpowerview.internal.api.responses.ScheduledEvents.ScheduledEvent;
-import org.openhab.binding.hdpowerview.internal.api.responses.Shade;
 import org.openhab.binding.hdpowerview.internal.api.responses.Shades;
-import org.openhab.binding.hdpowerview.internal.api.responses.Shades.ShadeData;
-import org.openhab.binding.hdpowerview.internal.api.responses.Survey;
-import org.openhab.binding.hdpowerview.internal.api.responses.UserDataResponse;
 import org.openhab.binding.hdpowerview.internal.exceptions.HubInvalidResponseException;
 import org.openhab.binding.hdpowerview.internal.exceptions.HubMaintenanceException;
 import org.openhab.binding.hdpowerview.internal.exceptions.HubProcessingException;
@@ -61,20 +50,20 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
-import com.google.gson.JsonParser;
 
 /**
- * JAX-RS targets for communicating with an HD PowerView hub
+ * Abstract class for JAX-RS targets for communicating with an HD PowerView hub.
  *
- * @author Andy Lintner - Initial contribution
- * @author Andrew Fiddian-Green - Added support for secondary rail positions
- * @author Jacob Laursen - Added support for scene groups and automations
+ * @author Andrew Fiddian-Green - Initial contribution
  */
 @NonNullByDefault
-public class HDPowerViewWebTargets {
+public abstract class HDPowerViewWebTargets {
 
     private final Logger logger = LoggerFactory.getLogger(HDPowerViewWebTargets.class);
 
@@ -87,24 +76,69 @@ public class HDPowerViewWebTargets {
     private final Duration maintenancePeriod = Duration.ofMinutes(5);
     private Instant maintenanceScheduledEnd = Instant.MIN;
 
-    private final String base;
-    private final String firmwareVersion;
-    private final String shades;
-    private final String sceneActivate;
-    private final String scenes;
-    private final String sceneCollectionActivate;
-    private final String sceneCollections;
-    private final String scheduledEvents;
-    private final String repeaters;
-    private final String userData;
+    /*
+     * De-serializer target class types
+     */
+    private final Class<?> sceneClass;
+    private final Class<?> shadeDataClass;
+    private final Class<?> shadePositionClass;
+    private final Class<?> scheduledEventClass;
 
-    private final Gson gson = new Gson();
-    private final HttpClient httpClient;
+    protected final Gson gson;
+    protected final HttpClient httpClient;
+
+    /*
+     * Private helper class for de-serialization of Scene
+     */
+    private class SceneDeserializer implements JsonDeserializer<Scene> {
+        @Override
+        public @Nullable Scene deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
+                throws JsonParseException {
+            JsonObject jsonObject = json.getAsJsonObject();
+            return context.deserialize(jsonObject, sceneClass);
+        }
+    }
+
+    /*
+     * Private helper class for de-serialization of ShadeData
+     */
+    private class ShadeDataDeserializer implements JsonDeserializer<ShadeData> {
+        @Override
+        public @Nullable ShadeData deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
+                throws JsonParseException {
+            JsonObject jsonObject = json.getAsJsonObject();
+            return context.deserialize(jsonObject, shadeDataClass);
+        }
+    }
+
+    /*
+     * Private helper class for de-serialization of ShadePosition
+     */
+    private class ShadePositionDeserializer implements JsonDeserializer<ShadePosition> {
+        @Override
+        public @Nullable ShadePosition deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
+                throws JsonParseException {
+            JsonObject jsonObject = json.getAsJsonObject();
+            return context.deserialize(jsonObject, shadePositionClass);
+        }
+    }
+
+    /*
+     * Private helper class for de-serialization of ScheduledEvent
+     */
+    private class ScheduledEventDeserializer implements JsonDeserializer<ScheduledEvent> {
+        @Override
+        public @Nullable ScheduledEvent deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
+                throws JsonParseException {
+            JsonObject jsonObject = json.getAsJsonObject();
+            return context.deserialize(jsonObject, scheduledEventClass);
+        }
+    }
 
     /**
-     * private helper class for passing http url query parameters
+     * Protected helper class for passing http url query parameters
      */
-    private static class Query {
+    protected static class Query {
         private final String key;
         private final String value;
 
@@ -137,429 +171,34 @@ public class HDPowerViewWebTargets {
      * @param httpClient the HTTP client (the binding)
      * @param ipAddress the IP address of the server (the hub)
      */
-    public HDPowerViewWebTargets(HttpClient httpClient, String ipAddress) {
-        base = "http://" + ipAddress + "/api/";
-        shades = base + "shades/";
-        firmwareVersion = base + "fwversion";
-        sceneActivate = base + "scenes";
-        scenes = base + "scenes/";
-
-        // Hub v1 only supports "scenecollections". Hub v2 will redirect to "sceneCollections".
-        sceneCollectionActivate = base + "scenecollections";
-        sceneCollections = base + "scenecollections/";
-
-        scheduledEvents = base + "scheduledevents";
-        repeaters = base + "repeaters/";
-        userData = base + "userdata";
-
+    public HDPowerViewWebTargets(HttpClient httpClient, String ipAddress, Class<?> sceneClass, Class<?> shadeDataClass,
+            Class<?> shadePositionClass, Class<?> scheduledEventClass) {
         this.httpClient = httpClient;
+        this.sceneClass = sceneClass;
+        this.shadeDataClass = shadeDataClass;
+        this.shadePositionClass = shadePositionClass;
+        this.scheduledEventClass = scheduledEventClass;
+        this.gson = new GsonBuilder()
+        // @formatter:off
+                .registerTypeAdapter(Scene.class, new SceneDeserializer())
+                .registerTypeAdapter(ShadeData.class, new ShadeDataDeserializer())
+                .registerTypeAdapter(ShadePosition.class, new ShadePositionDeserializer())
+                .registerTypeAdapter(ScheduledEvent.class, new ScheduledEventDeserializer())
+        // @formatter:on
+                .create();
     }
 
     /**
-     * Fetches a JSON package with firmware information for the hub.
+     * Get the gson de-serializer.
      *
-     * @return FirmwareVersions class instance
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
+     * @return the gson de-serializer.
      */
-    public HubFirmware getFirmwareVersions()
-            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException {
-        String json = invoke(HttpMethod.GET, firmwareVersion, null, null);
-        try {
-            FirmwareVersion firmwareVersion = gson.fromJson(json, FirmwareVersion.class);
-            if (firmwareVersion == null) {
-                throw new HubInvalidResponseException("Missing firmware response");
-            }
-            HubFirmware firmwareVersions = firmwareVersion.firmware;
-            if (firmwareVersions == null) {
-                throw new HubInvalidResponseException("Missing 'firmware' element");
-            }
-            return firmwareVersions;
-        } catch (JsonParseException e) {
-            throw new HubInvalidResponseException("Error parsing firmware response", e);
-        }
+    public Gson getGson() {
+        return gson;
     }
 
     /**
-     * Fetches a JSON package with user data information for the hub.
-     *
-     * @return {@link UserData} class instance
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     */
-    public UserData getUserData() throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException {
-        String json = invoke(HttpMethod.GET, userData, null, null);
-        try {
-            UserDataResponse userDataResponse = gson.fromJson(json, UserDataResponse.class);
-            if (userDataResponse == null) {
-                throw new HubInvalidResponseException("Missing userData response");
-            }
-            UserData userData = userDataResponse.userData;
-            if (userData == null) {
-                throw new HubInvalidResponseException("Missing 'userData' element");
-            }
-            return userData;
-        } catch (JsonParseException e) {
-            throw new HubInvalidResponseException("Error parsing userData response", e);
-        }
-    }
-
-    /**
-     * Fetches a JSON package that describes all shades in the hub, and wraps it in
-     * a Shades class instance
-     *
-     * @return Shades class instance
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     */
-    public Shades getShades() throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException {
-        String json = invoke(HttpMethod.GET, shades, null, null);
-        try {
-            Shades shades = gson.fromJson(json, Shades.class);
-            if (shades == null) {
-                throw new HubInvalidResponseException("Missing shades response");
-            }
-            List<ShadeData> shadeData = shades.shadeData;
-            if (shadeData == null) {
-                throw new HubInvalidResponseException("Missing 'shades.shadeData' element");
-            }
-            return shades;
-        } catch (JsonParseException e) {
-            throw new HubInvalidResponseException("Error parsing shades response", e);
-        }
-    }
-
-    /**
-     * Instructs the hub to move a specific shade
-     *
-     * @param shadeId id of the shade to be moved
-     * @param position instance of ShadePosition containing the new position
-     * @return ShadeData class instance (with new position)
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     * @throws HubShadeTimeoutException if the shade did not respond to a request
-     */
-    public ShadeData moveShade(int shadeId, ShadePosition position) throws HubInvalidResponseException,
-            HubProcessingException, HubMaintenanceException, HubShadeTimeoutException {
-        String jsonRequest = gson.toJson(new ShadeMove(position));
-        String jsonResponse = invoke(HttpMethod.PUT, shades + Integer.toString(shadeId), null, jsonRequest);
-        return shadeDataFromJson(jsonResponse);
-    }
-
-    private ShadeData shadeDataFromJson(String json) throws HubInvalidResponseException, HubShadeTimeoutException {
-        try {
-            Shade shade = gson.fromJson(json, Shade.class);
-            if (shade == null) {
-                throw new HubInvalidResponseException("Missing shade response");
-            }
-            ShadeData shadeData = shade.shade;
-            if (shadeData == null) {
-                throw new HubInvalidResponseException("Missing 'shade.shade' element");
-            }
-            if (Boolean.TRUE.equals(shadeData.timedOut)) {
-                throw new HubShadeTimeoutException("Timeout when sending request to the shade");
-            }
-            return shadeData;
-        } catch (JsonParseException e) {
-            throw new HubInvalidResponseException("Error parsing shade response", e);
-        }
-    }
-
-    /**
-     * Instructs the hub to stop movement of a specific shade
-     *
-     * @param shadeId id of the shade to be stopped
-     * @return ShadeData class instance (new position cannot be relied upon)
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     * @throws HubShadeTimeoutException if the shade did not respond to a request
-     */
-    public ShadeData stopShade(int shadeId) throws HubInvalidResponseException, HubProcessingException,
-            HubMaintenanceException, HubShadeTimeoutException {
-        String jsonRequest = gson.toJson(new ShadeStop());
-        String jsonResponse = invoke(HttpMethod.PUT, shades + Integer.toString(shadeId), null, jsonRequest);
-        return shadeDataFromJson(jsonResponse);
-    }
-
-    /**
-     * Instructs the hub to jog a specific shade
-     *
-     * @param shadeId id of the shade to be jogged
-     * @return ShadeData class instance
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     * @throws HubShadeTimeoutException if the shade did not respond to a request
-     */
-    public ShadeData jogShade(int shadeId) throws HubInvalidResponseException, HubProcessingException,
-            HubMaintenanceException, HubShadeTimeoutException {
-        String jsonRequest = gson.toJson(new ShadeJog());
-        String jsonResponse = invoke(HttpMethod.PUT, shades + Integer.toString(shadeId), null, jsonRequest);
-        return shadeDataFromJson(jsonResponse);
-    }
-
-    /**
-     * Instructs the hub to calibrate a specific shade
-     *
-     * @param shadeId id of the shade to be calibrated
-     * @return ShadeData class instance
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     * @throws HubShadeTimeoutException if the shade did not respond to a request
-     */
-    public ShadeData calibrateShade(int shadeId) throws HubInvalidResponseException, HubProcessingException,
-            HubMaintenanceException, HubShadeTimeoutException {
-        String jsonRequest = gson.toJson(new ShadeCalibrate());
-        String jsonResponse = invoke(HttpMethod.PUT, shades + Integer.toString(shadeId), null, jsonRequest);
-        return shadeDataFromJson(jsonResponse);
-    }
-
-    /**
-     * Fetches a JSON package that describes all scenes in the hub, and wraps it in
-     * a Scenes class instance
-     *
-     * @return Scenes class instance
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     */
-    public Scenes getScenes() throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException {
-        String json = invoke(HttpMethod.GET, scenes, null, null);
-        try {
-            Scenes scenes = gson.fromJson(json, Scenes.class);
-            if (scenes == null) {
-                throw new HubInvalidResponseException("Missing scenes response");
-            }
-            List<Scene> sceneData = scenes.sceneData;
-            if (sceneData == null) {
-                throw new HubInvalidResponseException("Missing 'scenes.sceneData' element");
-            }
-            return scenes;
-        } catch (JsonParseException e) {
-            throw new HubInvalidResponseException("Error parsing scenes response", e);
-        }
-    }
-
-    /**
-     * Instructs the hub to execute a specific scene
-     *
-     * @param sceneId id of the scene to be executed
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     */
-    public void activateScene(int sceneId) throws HubProcessingException, HubMaintenanceException {
-        invoke(HttpMethod.GET, sceneActivate, Query.of("sceneId", Integer.toString(sceneId)), null);
-    }
-
-    /**
-     * Fetches a JSON package that describes all scene collections in the hub, and wraps it in
-     * a SceneCollections class instance
-     *
-     * @return SceneCollections class instance
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     */
-    public SceneCollections getSceneCollections()
-            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException {
-        String json = invoke(HttpMethod.GET, sceneCollections, null, null);
-        try {
-            SceneCollections sceneCollections = gson.fromJson(json, SceneCollections.class);
-            if (sceneCollections == null) {
-                throw new HubInvalidResponseException("Missing sceneCollections response");
-            }
-            List<SceneCollection> sceneCollectionData = sceneCollections.sceneCollectionData;
-            if (sceneCollectionData == null) {
-                throw new HubInvalidResponseException("Missing 'sceneCollections.sceneCollectionData' element");
-            }
-            return sceneCollections;
-        } catch (JsonParseException e) {
-            throw new HubInvalidResponseException("Error parsing sceneCollections response", e);
-        }
-    }
-
-    /**
-     * Instructs the hub to execute a specific scene collection
-     *
-     * @param sceneCollectionId id of the scene collection to be executed
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     */
-    public void activateSceneCollection(int sceneCollectionId) throws HubProcessingException, HubMaintenanceException {
-        invoke(HttpMethod.GET, sceneCollectionActivate,
-                Query.of("sceneCollectionId", Integer.toString(sceneCollectionId)), null);
-    }
-
-    /**
-     * Fetches a JSON package that describes all scheduled events in the hub, and wraps it in
-     * a ScheduledEvents class instance
-     *
-     * @return ScheduledEvents class instance
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     */
-    public ScheduledEvents getScheduledEvents()
-            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException {
-        String json = invoke(HttpMethod.GET, scheduledEvents, null, null);
-        try {
-            ScheduledEvents scheduledEvents = gson.fromJson(json, ScheduledEvents.class);
-            if (scheduledEvents == null) {
-                throw new HubInvalidResponseException("Missing scheduledEvents response");
-            }
-            List<ScheduledEvent> scheduledEventData = scheduledEvents.scheduledEventData;
-            if (scheduledEventData == null) {
-                throw new HubInvalidResponseException("Missing 'scheduledEvents.scheduledEventData' element");
-            }
-            return scheduledEvents;
-        } catch (JsonParseException e) {
-            throw new HubInvalidResponseException("Error parsing scheduledEvents response", e);
-        }
-    }
-
-    /**
-     * Enables or disables a scheduled event in the hub.
-     * 
-     * @param scheduledEventId id of the scheduled event to be enabled or disabled
-     * @param enable true to enable scheduled event, false to disable
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     */
-    public void enableScheduledEvent(int scheduledEventId, boolean enable)
-            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException {
-        String uri = scheduledEvents + "/" + scheduledEventId;
-        String jsonResponse = invoke(HttpMethod.GET, uri, null, null);
-        try {
-            JsonObject jsonObject = JsonParser.parseString(jsonResponse).getAsJsonObject();
-            JsonElement scheduledEventElement = jsonObject.get("scheduledEvent");
-            if (scheduledEventElement == null) {
-                throw new HubInvalidResponseException("Missing 'scheduledEvent' element");
-            }
-            JsonObject scheduledEventObject = scheduledEventElement.getAsJsonObject();
-            scheduledEventObject.addProperty("enabled", enable);
-            invoke(HttpMethod.PUT, uri, null, jsonObject.toString());
-        } catch (JsonParseException | IllegalStateException e) {
-            throw new HubInvalidResponseException("Error parsing scheduledEvent response", e);
-        }
-    }
-
-    /**
-     * Fetches a JSON package that describes all repeaters in the hub, and wraps it in
-     * a Repeaters class instance
-     *
-     * @return Repeaters class instance
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     */
-    public Repeaters getRepeaters()
-            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException {
-        String json = invoke(HttpMethod.GET, repeaters, null, null);
-        try {
-            Repeaters repeaters = gson.fromJson(json, Repeaters.class);
-            if (repeaters == null) {
-                throw new HubInvalidResponseException("Missing repeaters response");
-            }
-            List<RepeaterData> repeaterData = repeaters.repeaterData;
-            if (repeaterData == null) {
-                throw new HubInvalidResponseException("Missing 'repeaters.repeaterData' element");
-            }
-            return repeaters;
-        } catch (JsonParseException e) {
-            throw new HubInvalidResponseException("Error parsing repeaters response", e);
-        }
-    }
-
-    /**
-     * Fetches a JSON package that describes a specific repeater in the hub, and wraps it
-     * in a RepeaterData class instance
-     *
-     * @param repeaterId id of the repeater to be fetched
-     * @return RepeaterData class instance
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     */
-    public RepeaterData getRepeater(int repeaterId)
-            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException {
-        String jsonResponse = invoke(HttpMethod.GET, repeaters + Integer.toString(repeaterId), null, null);
-        return repeaterDataFromJson(jsonResponse);
-    }
-
-    private RepeaterData repeaterDataFromJson(String json) throws HubInvalidResponseException {
-        try {
-            Repeater repeater = gson.fromJson(json, Repeater.class);
-            if (repeater == null) {
-                throw new HubInvalidResponseException("Missing repeater response");
-            }
-            RepeaterData repeaterData = repeater.repeater;
-            if (repeaterData == null) {
-                throw new HubInvalidResponseException("Missing 'repeater.repeater' element");
-            }
-            return repeaterData;
-        } catch (JsonParseException e) {
-            throw new HubInvalidResponseException("Error parsing repeater response", e);
-        }
-    }
-
-    /**
-     * Instructs the hub to identify a specific repeater by blinking
-     *
-     * @param repeaterId id of the repeater to be identified
-     * @return RepeaterData class instance
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     */
-    public RepeaterData identifyRepeater(int repeaterId)
-            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException {
-        String jsonResponse = invoke(HttpMethod.GET, repeaters + repeaterId,
-                Query.of("identify", Boolean.toString(true)), null);
-        return repeaterDataFromJson(jsonResponse);
-    }
-
-    /**
-     * Enables or disables blinking for a repeater
-     * 
-     * @param repeaterId id of the repeater for which to be enable or disable blinking
-     * @param enable true to enable blinking, false to disable
-     * @return RepeaterData class instance
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     */
-    public RepeaterData enableRepeaterBlinking(int repeaterId, boolean enable)
-            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException {
-        String jsonRequest = gson.toJson(new RepeaterBlinking(repeaterId, enable));
-        String jsonResponse = invoke(HttpMethod.PUT, repeaters + repeaterId, null, jsonRequest);
-        return repeaterDataFromJson(jsonResponse);
-    }
-
-    /**
-     * Sets color and brightness for a repeater
-     *
-     * @param repeaterId id of the repeater for which to set color and brightness
-     * @return RepeaterData class instance
-     * @throws HubInvalidResponseException if response is invalid
-     * @throws HubProcessingException if there is any processing error
-     * @throws HubMaintenanceException if the hub is down for maintenance
-     */
-    public RepeaterData setRepeaterColor(int repeaterId, Color color)
-            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException {
-        String jsonRequest = gson.toJson(new RepeaterColor(repeaterId, color));
-        String jsonResponse = invoke(HttpMethod.PUT, repeaters + repeaterId, null, jsonRequest);
-        return repeaterDataFromJson(jsonResponse);
-    }
-
-    /**
-     * Invoke a call on the hub server to retrieve information or send a command
+     * Common protected method to invoke a call on the hub server to retrieve information or send a command
      *
      * @param method GET or PUT
      * @param url the host url to be called
@@ -569,7 +208,7 @@ public class HDPowerViewWebTargets {
      * @throws HubMaintenanceException
      * @throws HubProcessingException
      */
-    private synchronized String invoke(HttpMethod method, String url, @Nullable Query query,
+    protected synchronized String invoke(HttpMethod method, String url, @Nullable Query query,
             @Nullable String jsonCommand) throws HubMaintenanceException, HubProcessingException {
         if (logger.isTraceEnabled()) {
             if (query != null) {
@@ -625,6 +264,222 @@ public class HDPowerViewWebTargets {
     }
 
     /**
+     * Fetches a JSON package that describes all shades in the hub, and wraps it in
+     * a Shades class instance
+     *
+     * @return Shades class instance
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract Shades getShades()
+            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException;
+
+    /**
+     * Fetches a JSON package that describes all scenes in the hub, and wraps it in
+     * a Scenes class instance
+     *
+     * @return Scenes class instance
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract Scenes getScenes()
+            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException;
+
+    /**
+     * Fetches a JSON package with firmware information for the hub.
+     *
+     * @return FirmwareVersions class instance
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract HubFirmware getFirmwareVersions()
+            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException;
+
+    /**
+     * Fetches a JSON package with user data information for the hub.
+     *
+     * @return {@link UserData} class instance
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract UserData getUserData()
+            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException;
+
+    /**
+     * Instructs the hub to move a specific shade
+     *
+     * @param shadeId id of the shade to be moved
+     * @param position instance of ShadePosition containing the new position
+     * @return ShadeData class instance (with new position)
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     * @throws HubShadeTimeoutException if the shade did not respond to a request
+     */
+    public abstract ShadeData moveShade(int shadeId, ShadePosition position) throws HubInvalidResponseException,
+            HubProcessingException, HubMaintenanceException, HubShadeTimeoutException;
+
+    /**
+     * Instructs the hub to stop movement of a specific shade
+     *
+     * @param shadeId id of the shade to be stopped
+     * @return ShadeData class instance (new position cannot be relied upon)
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     * @throws HubShadeTimeoutException if the shade did not respond to a request
+     */
+    public abstract ShadeData stopShade(int shadeId) throws HubInvalidResponseException, HubProcessingException,
+            HubMaintenanceException, HubShadeTimeoutException;
+
+    /**
+     * Instructs the hub to jog a specific shade
+     *
+     * @param shadeId id of the shade to be jogged
+     * @return ShadeData class instance
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     * @throws HubShadeTimeoutException if the shade did not respond to a request
+     */
+    public abstract ShadeData jogShade(int shadeId) throws HubInvalidResponseException, HubProcessingException,
+            HubMaintenanceException, HubShadeTimeoutException;
+
+    /**
+     * Instructs the hub to calibrate a specific shade
+     *
+     * @param shadeId id of the shade to be calibrated
+     * @return ShadeData class instance
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     * @throws HubShadeTimeoutException if the shade did not respond to a request
+     */
+    public abstract ShadeData calibrateShade(int shadeId) throws HubInvalidResponseException, HubProcessingException,
+            HubMaintenanceException, HubShadeTimeoutException;
+
+    /**
+     * Instructs the hub to execute a specific scene
+     *
+     * @param sceneId id of the scene to be executed
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract void activateScene(int sceneId) throws HubProcessingException, HubMaintenanceException;
+
+    /**
+     * Fetches a JSON package that describes all scene collections in the hub, and wraps it in
+     * a SceneCollections class instance
+     *
+     * @return SceneCollections class instance
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract SceneCollections getSceneCollections()
+            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException;
+
+    /**
+     * Instructs the hub to execute a specific scene collection
+     *
+     * @param sceneCollectionId id of the scene collection to be executed
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract void activateSceneCollection(int sceneCollectionId)
+            throws HubProcessingException, HubMaintenanceException;
+
+    /**
+     * Fetches a JSON package that describes all scheduled events in the hub, and wraps it in
+     * a ScheduledEvents class instance
+     *
+     * @return ScheduledEvents class instance
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract ScheduledEvents getScheduledEvents()
+            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException;
+
+    /**
+     * Enables or disables a scheduled event in the hub.
+     *
+     * @param scheduledEventId id of the scheduled event to be enabled or disabled
+     * @param enable true to enable scheduled event, false to disable
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract void enableScheduledEvent(int scheduledEventId, boolean enable)
+            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException;
+
+    /**
+     * Fetches a JSON package that describes all repeaters in the hub, and wraps it in
+     * a Repeaters class instance
+     *
+     * @return Repeaters class instance
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract Repeaters getRepeaters()
+            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException;
+
+    /**
+     * Fetches a JSON package that describes a specific repeater in the hub, and wraps it
+     * in a RepeaterData class instance
+     *
+     * @param repeaterId id of the repeater to be fetched
+     * @return RepeaterData class instance
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract RepeaterData getRepeater(int repeaterId)
+            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException;
+
+    /**
+     * Instructs the hub to identify a specific repeater by blinking
+     *
+     * @param repeaterId id of the repeater to be identified
+     * @return RepeaterData class instance
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract RepeaterData identifyRepeater(int repeaterId)
+            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException;
+
+    /**
+     * Enables or disables blinking for a repeater
+     *
+     * @param repeaterId id of the repeater for which to be enable or disable blinking
+     * @param enable true to enable blinking, false to disable
+     * @return RepeaterData class instance
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract RepeaterData enableRepeaterBlinking(int repeaterId, boolean enable)
+            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException;
+
+    /**
+     * Sets color and brightness for a repeater
+     *
+     * @param repeaterId id of the repeater for which to set color and brightness
+     * @return RepeaterData class instance
+     * @throws HubInvalidResponseException if response is invalid
+     * @throws HubProcessingException if there is any processing error
+     * @throws HubMaintenanceException if the hub is down for maintenance
+     */
+    public abstract RepeaterData setRepeaterColor(int repeaterId, Color color)
+            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException;
+
+    /**
      * Fetches a JSON package that describes a specific shade in the hub, and wraps it
      * in a Shade class instance
      *
@@ -635,11 +490,8 @@ public class HDPowerViewWebTargets {
      * @throws HubMaintenanceException if the hub is down for maintenance
      * @throws HubShadeTimeoutException if the shade did not respond to a request
      */
-    public ShadeData getShade(int shadeId) throws HubInvalidResponseException, HubProcessingException,
-            HubMaintenanceException, HubShadeTimeoutException {
-        String jsonResponse = invoke(HttpMethod.GET, shades + Integer.toString(shadeId), null, null);
-        return shadeDataFromJson(jsonResponse);
-    }
+    public abstract ShadeData getShade(int shadeId) throws HubInvalidResponseException, HubProcessingException,
+            HubMaintenanceException, HubShadeTimeoutException;
 
     /**
      * Instructs the hub to do a hard refresh (discovery on the hubs RF network) on
@@ -653,12 +505,8 @@ public class HDPowerViewWebTargets {
      * @throws HubMaintenanceException if the hub is down for maintenance
      * @throws HubShadeTimeoutException if the shade did not respond to a request
      */
-    public ShadeData refreshShadePosition(int shadeId)
-            throws JsonParseException, HubProcessingException, HubMaintenanceException, HubShadeTimeoutException {
-        String jsonResponse = invoke(HttpMethod.GET, shades + Integer.toString(shadeId),
-                Query.of("refresh", Boolean.toString(true)), null);
-        return shadeDataFromJson(jsonResponse);
-    }
+    public abstract ShadeData refreshShadePosition(int shadeId)
+            throws JsonParseException, HubProcessingException, HubMaintenanceException, HubShadeTimeoutException;
 
     /**
      * Instructs the hub to do a hard refresh (discovery on the hubs RF network) on
@@ -672,24 +520,8 @@ public class HDPowerViewWebTargets {
      * @throws HubProcessingException if there is any processing error
      * @throws HubMaintenanceException if the hub is down for maintenance
      */
-    public List<SurveyData> getShadeSurvey(int shadeId)
-            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException {
-        String jsonResponse = invoke(HttpMethod.GET, shades + Integer.toString(shadeId),
-                Query.of("survey", Boolean.toString(true)), null);
-        try {
-            Survey survey = gson.fromJson(jsonResponse, Survey.class);
-            if (survey == null) {
-                throw new HubInvalidResponseException("Missing survey response");
-            }
-            List<SurveyData> surveyData = survey.surveyData;
-            if (surveyData == null) {
-                throw new HubInvalidResponseException("Missing 'survey.surveyData' element");
-            }
-            return surveyData;
-        } catch (JsonParseException e) {
-            throw new HubInvalidResponseException("Error parsing survey response", e);
-        }
-    }
+    public abstract List<SurveyData> getShadeSurvey(int shadeId)
+            throws HubInvalidResponseException, HubProcessingException, HubMaintenanceException;
 
     /**
      * Instructs the hub to do a hard refresh (discovery on the hubs RF network) on
@@ -703,10 +535,6 @@ public class HDPowerViewWebTargets {
      * @throws HubMaintenanceException if the hub is down for maintenance
      * @throws HubShadeTimeoutException if the shade did not respond to a request
      */
-    public ShadeData refreshShadeBatteryLevel(int shadeId) throws HubInvalidResponseException, HubProcessingException,
-            HubMaintenanceException, HubShadeTimeoutException {
-        String jsonResponse = invoke(HttpMethod.GET, shades + Integer.toString(shadeId),
-                Query.of("updateBatteryLevel", Boolean.toString(true)), null);
-        return shadeDataFromJson(jsonResponse);
-    }
+    public abstract ShadeData refreshShadeBatteryLevel(int shadeId) throws HubInvalidResponseException,
+            HubProcessingException, HubMaintenanceException, HubShadeTimeoutException;
 }
