+++ bundles/org.openhab.binding.hdpowerview/src/main/java/org/openhab/binding/hdpowerview/internal/handler/HDPowerViewHubHandler.java
@@ -30,17 +30,19 @@ import org.eclipse.jetty.client.HttpClient;
 import org.openhab.binding.hdpowerview.internal.HDPowerViewBindingConstants;
 import org.openhab.binding.hdpowerview.internal.HDPowerViewTranslationProvider;
 import org.openhab.binding.hdpowerview.internal.HDPowerViewWebTargets;
+import org.openhab.binding.hdpowerview.internal._v1.HDPowerViewWebTargetsV1;
 import org.openhab.binding.hdpowerview.internal.api.Firmware;
 import org.openhab.binding.hdpowerview.internal.api.HubFirmware;
+import org.openhab.binding.hdpowerview.internal.api.ShadeData;
 import org.openhab.binding.hdpowerview.internal.api.UserData;
+import org.openhab.binding.hdpowerview.internal.api._v1.ShadeDataV1;
+import org.openhab.binding.hdpowerview.internal.api.responses.Scene;
 import org.openhab.binding.hdpowerview.internal.api.responses.SceneCollections;
 import org.openhab.binding.hdpowerview.internal.api.responses.SceneCollections.SceneCollection;
 import org.openhab.binding.hdpowerview.internal.api.responses.Scenes;
-import org.openhab.binding.hdpowerview.internal.api.responses.Scenes.Scene;
+import org.openhab.binding.hdpowerview.internal.api.responses.ScheduledEvent;
 import org.openhab.binding.hdpowerview.internal.api.responses.ScheduledEvents;
-import org.openhab.binding.hdpowerview.internal.api.responses.ScheduledEvents.ScheduledEvent;
 import org.openhab.binding.hdpowerview.internal.api.responses.Shades;
-import org.openhab.binding.hdpowerview.internal.api.responses.Shades.ShadeData;
 import org.openhab.binding.hdpowerview.internal.builders.AutomationChannelBuilder;
 import org.openhab.binding.hdpowerview.internal.builders.SceneChannelBuilder;
 import org.openhab.binding.hdpowerview.internal.builders.SceneGroupChannelBuilder;
@@ -101,6 +103,9 @@ public class HDPowerViewHubHandler extends BaseBridgeHandler {
     private List<ScheduledEvent> scheduledEventCache = new CopyOnWriteArrayList<>();
     private Instant userDataUpdated = Instant.MIN;
     private Boolean deprecatedChannelsCreated = false;
+
+    // TODO auto initialize this field or via config parameter
+    // private boolean isGeneration1 = true;
 
     private final ChannelTypeUID sceneChannelTypeUID = new ChannelTypeUID(HDPowerViewBindingConstants.BINDING_ID,
             HDPowerViewBindingConstants.CHANNELTYPE_SCENE_ACTIVATE);
@@ -164,7 +169,7 @@ public class HDPowerViewHubHandler extends BaseBridgeHandler {
         }
 
         pendingShadeInitializations.clear();
-        webTargets = new HDPowerViewWebTargets(httpClient, host);
+        webTargets = newWebTargets(host);
         refreshInterval = config.refresh;
         hardRefreshPositionInterval = config.hardRefresh;
         hardRefreshBatteryLevelInterval = config.hardRefreshBatteryLevel;
@@ -422,7 +427,7 @@ public class HDPowerViewHubHandler extends BaseBridgeHandler {
         HDPowerViewShadeHandler thingHandler = ((HDPowerViewShadeHandler) thing.getHandler());
         if (thingHandler == null) {
             logger.debug("Shade '{}' handler not initialized", shadeId);
-            pendingShadeInitializations.put(thing.getUID(), new ShadeData());
+            pendingShadeInitializations.put(thing.getUID(), newShadeData());
             return;
         }
         ThingStatus thingStatus = thingHandler.getThing().getStatus();
@@ -436,7 +441,7 @@ public class HDPowerViewHubHandler extends BaseBridgeHandler {
             case UNINITIALIZED:
             case INITIALIZING:
                 logger.debug("Shade '{}' handler not yet ready; status: {}", shadeId, thingStatus);
-                pendingShadeInitializations.put(thing.getUID(), new ShadeData());
+                pendingShadeInitializations.put(thing.getUID(), newShadeData());
                 break;
             case REMOVING:
             case REMOVED:
@@ -674,5 +679,16 @@ public class HDPowerViewHubHandler extends BaseBridgeHandler {
                 logger.debug("Shade '{}' handler not initialized", shadeId);
             }
         }
+    }
+
+    private HDPowerViewWebTargets newWebTargets(String host) {
+        // return isGeneration1 ? new HDPowerViewWebTargetsV1(httpClient, host)
+        // : new HDPowerViewWebTargetsV3(httpClient, host);
+        return new HDPowerViewWebTargetsV1(httpClient, host);
+    }
+
+    private ShadeData newShadeData() {
+        // return isGeneration1 ? new ShadeDataV1() : new ShadeDataV3();
+        return new ShadeDataV1();
     }
 }
