+++ bundles/org.openhab.binding.hdpowerview/src/main/java/org/openhab/binding/hdpowerview/internal/handler/HDPowerViewShadeHandler.java
@@ -32,9 +32,11 @@ import org.openhab.binding.hdpowerview.internal.HDPowerViewWebTargets;
 import org.openhab.binding.hdpowerview.internal.api.BatteryKind;
 import org.openhab.binding.hdpowerview.internal.api.CoordinateSystem;
 import org.openhab.binding.hdpowerview.internal.api.Firmware;
+import org.openhab.binding.hdpowerview.internal.api.ShadeData;
 import org.openhab.binding.hdpowerview.internal.api.ShadePosition;
 import org.openhab.binding.hdpowerview.internal.api.SurveyData;
-import org.openhab.binding.hdpowerview.internal.api.responses.Shades.ShadeData;
+import org.openhab.binding.hdpowerview.internal.api._v1.ShadeDataV1;
+import org.openhab.binding.hdpowerview.internal.api._v1.ShadePositionV1;
 import org.openhab.binding.hdpowerview.internal.config.HDPowerViewShadeConfiguration;
 import org.openhab.binding.hdpowerview.internal.database.ShadeCapabilitiesDatabase;
 import org.openhab.binding.hdpowerview.internal.database.ShadeCapabilitiesDatabase.Capabilities;
@@ -94,6 +96,10 @@ public class HDPowerViewShadeHandler extends AbstractHubbedThingHandler {
     private @Nullable Capabilities capabilities;
     private int shadeId;
     private boolean isDisposing;
+
+    // private boolean isGeneration1 = true;
+
+    // private boolean isGeneration1 = true;
 
     public HDPowerViewShadeHandler(Thing thing) {
         super(thing);
@@ -250,6 +256,7 @@ public class HDPowerViewShadeHandler extends AbstractHubbedThingHandler {
      * @param shadeData the ShadeData to be used.
      */
     protected void onReceiveUpdate(ShadeData shadeData) {
+        // isGeneration1 = shadeData.version() == 1;
         updateStatus(ThingStatus.ONLINE);
         updateCapabilities(shadeData);
         updateSoftProperties(shadeData);
@@ -258,7 +265,7 @@ public class HDPowerViewShadeHandler extends AbstractHubbedThingHandler {
         if (shadePosition != null) {
             updatePositionStates(shadePosition);
         }
-        updateBatteryStates(shadeData.batteryStatus, shadeData.batteryStrength);
+        updateBatteryStates(shadeData);
         updateSignalStrengthState(shadeData.signalStrength);
     }
 
@@ -334,7 +341,7 @@ public class HDPowerViewShadeHandler extends AbstractHubbedThingHandler {
     private void updateFirmwareProperties(ShadeData shadeData) {
         Map<String, String> properties = editProperties();
         Firmware shadeFirmware = shadeData.firmware;
-        Firmware motorFirmware = shadeData.motor;
+        Firmware motorFirmware = (shadeData.version() == 1) ? ((ShadeDataV1) shadeData).motor : null;
         if (shadeFirmware != null) {
             properties.put(Thing.PROPERTY_FIRMWARE_VERSION, shadeFirmware.toString());
         }
@@ -399,8 +406,9 @@ public class HDPowerViewShadeHandler extends AbstractHubbedThingHandler {
         updateState(CHANNEL_SHADE_SECONDARY_POSITION, shadePos.getState(capabilities, SECONDARY_POSITION));
     }
 
-    private void updateBatteryStates(int batteryStatus, double batteryStrength) {
-        updateBatteryLevelStates(batteryStatus);
+    private void updateBatteryStates(ShadeData shadeData) {
+        updateBatteryLevelStates(shadeData.batteryStatus);
+        double batteryStrength = shadeData.version() == 1 ? ((ShadeDataV1) shadeData).batteryStrength : 0;
         updateState(CHANNEL_SHADE_BATTERY_VOLTAGE,
                 batteryStrength > 0 ? new QuantityType<>(batteryStrength / 10, Units.VOLT) : UnDefType.UNDEF);
     }
@@ -441,7 +449,7 @@ public class HDPowerViewShadeHandler extends AbstractHubbedThingHandler {
         newPosition = shadeData.positions;
         // if no positions returned, then create a new position
         if (newPosition == null) {
-            newPosition = new ShadePosition();
+            newPosition = newShadePosition();
         }
         Capabilities capabilities = getCapabilitiesOrDefault();
         // set the new position value, and write the positions to the hub
@@ -570,7 +578,7 @@ public class HDPowerViewShadeHandler extends AbstractHubbedThingHandler {
                     break;
                 case BATTERY_LEVEL:
                     shadeData = webTargets.refreshShadeBatteryLevel(shadeId);
-                    updateBatteryStates(shadeData.batteryStatus, shadeData.batteryStrength);
+                    updateBatteryStates(shadeData);
                     break;
                 default:
                     throw new NotSupportedException("Unsupported refresh kind " + kind.toString());
@@ -647,5 +655,10 @@ public class HDPowerViewShadeHandler extends AbstractHubbedThingHandler {
             }
             updateThing(editThing().withoutChannels(removeList).build());
         }
+    }
+
+    private ShadePosition newShadePosition() {
+        // return isGeneration1 ? new ShadePositionV1() : new ShadePositionV3();
+        return new ShadePositionV1();
     }
 }
