+++ bundles/org.openhab.binding.hdpowerview/src/test/java/org/openhab/binding/hdpowerview/OnlineCommunicationTest.java
@@ -13,7 +13,7 @@
 package org.openhab.binding.hdpowerview;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.openhab.binding.hdpowerview.internal.api.CoordinateSystem.*;
+import static org.openhab.binding.hdpowerview.internal.api.CoordinateSystem.PRIMARY_POSITION;
 
 import java.util.List;
 import java.util.regex.Pattern;
@@ -22,11 +22,13 @@ import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jetty.client.HttpClient;
 import org.junit.jupiter.api.Test;
 import org.openhab.binding.hdpowerview.internal.HDPowerViewWebTargets;
+import org.openhab.binding.hdpowerview.internal._v1.HDPowerViewWebTargetsV1;
+import org.openhab.binding.hdpowerview.internal.api.ShadeData;
 import org.openhab.binding.hdpowerview.internal.api.ShadePosition;
+import org.openhab.binding.hdpowerview.internal.api._v1.ShadePositionV1;
+import org.openhab.binding.hdpowerview.internal.api.responses.Scene;
 import org.openhab.binding.hdpowerview.internal.api.responses.Scenes;
-import org.openhab.binding.hdpowerview.internal.api.responses.Scenes.Scene;
 import org.openhab.binding.hdpowerview.internal.api.responses.Shades;
-import org.openhab.binding.hdpowerview.internal.api.responses.Shades.ShadeData;
 import org.openhab.binding.hdpowerview.internal.database.ShadeCapabilitiesDatabase;
 import org.openhab.binding.hdpowerview.internal.database.ShadeCapabilitiesDatabase.Capabilities;
 import org.openhab.binding.hdpowerview.internal.exceptions.HubException;
@@ -83,7 +85,7 @@ public class OnlineCommunicationTest {
                 fail(e.getMessage());
             }
 
-            HDPowerViewWebTargets webTargets = new HDPowerViewWebTargets(client, hubIPAddress);
+            HDPowerViewWebTargets webTargets = new HDPowerViewWebTargetsV1(client, hubIPAddress);
             assertNotNull(webTargets);
 
             int shadeId = 0;
@@ -168,7 +170,7 @@ public class OnlineCommunicationTest {
                     int position = ((PercentType) pos).intValue();
                     position = position + ((position <= 10) ? 5 : -5);
 
-                    ShadePosition targetPosition = new ShadePosition().setPosition(capabilities, PRIMARY_POSITION,
+                    ShadePosition targetPosition = new ShadePositionV1().setPosition(capabilities, PRIMARY_POSITION,
                             position);
                     assertNotNull(targetPosition);
